Move field types to extension.ts file
