--- src/lib/extension.ts
@@ -1,6 +1,0 @@
-export const IDENTIFIER = "aha-develop.github";
-export const ACTIONS_IDENTIFIER = "aha-develop.github-actions";
-
-export const LEARN_MORE_URL =
-  "https://www.aha.io/support/develop/integrations/github/github-extension";
-export const ICON = aha.iconForExtensionIdentifier(IDENTIFIER);
