+++ src/views/prPanel.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { ExtensionRoot } from "@components/ExtensionRoot";
 import { PrPanel } from "@components/PrPanel";
 import { TableCols } from "@components/PrTable/PrTable";
-import { IDENTIFIER } from "@lib/extension";
+import { IDENTIFIER } from "extension";
 
 const panel = aha.getPanel(IDENTIFIER, "prPanel", {
   name: "GitHub pull requests",
